feat: comment 관련 type을 한곳에서 관리한다
